COMMON: Utility functions for converting a TimeDate into an integer and vice versa
