+++ engines/twp/metaengine.cpp
@@ -37,7 +37,6 @@
 #include "twp/metaengine.h"
 #include "twp/detection.h"
 #include "twp/savegame.h"
-#include "twp/time.h"
 #include "twp/actions.h"
 #include "twp/debugtools.h"
 #include "twp/dialogs.h"
@@ -83,7 +82,7 @@ const Common::AchievementDescriptionList *TwpMetaEngine::getAchievementDescripti
 }
 
 static Common::String getDesc(const Twp::SaveGame &savegame) {
-	Common::String desc = Twp::formatTime(savegame.time);
+	Common::String desc = Common::DateTime::formatTime(savegame.time);
 	if (savegame.easyMode)
 		desc += " (casual)";
 	return desc;
@@ -115,7 +114,7 @@ SaveStateDescriptor TwpMetaEngine::querySaveMetaInfos(const char *target, int sl
 		Twp::SaveGame savegame;
 		Twp::SaveGameManager::getSaveGame(f, savegame);
 		Common::String savegameDesc(getDesc(savegame));
-		TimeDate dt = Twp::intToTimeDate(savegame.time);
+		TimeDate dt = Common::DateTime::int64ToTimeDate(savegame.time);
 		desc.setDescription(savegameDesc);
 		desc.setPlayTime(savegame.gameTime * 1000);
 		desc.setSaveDate(dt.tm_year, dt.tm_mon, dt.tm_mday);
